+++ sources/nix/stdio/vsnprintf.c
@@ -1,10 +1,10 @@
-#include <stdio.h>
 #include <stdarg.h>
+#include "stdio.h"
 
 extern int __vfprintf_total_size(FILE *stream, const char *fmt, va_list args);
 
 int vsnprintf(char *s,size_t size,const char *format,va_list args) {
-	int retval;
+	size_t retval;
 	FILE buffer;
 
 	if (!s && size )
